refactored waitlist api to be a part of camper api
