+++ backend/typescript/server.ts
@@ -11,7 +11,6 @@ import entityRouter from "./rest/entityRoutes";
 import userRouter from "./rest/userRoutes";
 import campRouter from "./rest/campRoutes";
 import camperRouter from "./rest/camperRoutes";
-import waitlistRouter from "./rest/waitlistRoutes";
 
 const CORS_ALLOW_LIST = [
   "http://localhost:3000",
@@ -38,7 +37,6 @@ app.use("/entities", entityRouter);
 app.use("/users", userRouter);
 app.use("/camp", campRouter);
 app.use("/camper", camperRouter);
-app.use("/waitlist", waitlistRouter);
 app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 
 mongo.connect();
