+++ backend/typescript/services/implementations/camperService.ts
@@ -1,7 +1,15 @@
 import ICamperService from "../interfaces/camperService";
 import MgCamper, { Camper } from "../../models/camper.model";
+import MgWaitlistedCamper, {
+  WaitlistedCamper,
+} from "../../models/waitlistedCamper.model";
 import MgCamp, { Camp } from "../../models/camp.model";
-import { CreateCamperDTO, CamperDTO } from "../../types";
+import {
+  CreateCamperDTO,
+  CamperDTO,
+  CreateWaitlistedCamperDTO,
+  WaitlistedCamperDTO,
+} from "../../types";
 import { getErrorMessage } from "../../utilities/errorUtils";
 import logger from "../../utilities/logger";
 
@@ -166,6 +174,78 @@ class CamperService implements ICamperService {
     }
 
     return camperDtos;
+  }
+
+  async createWaitlistedCamper(
+    waitlistedCamper: CreateWaitlistedCamperDTO,
+  ): Promise<WaitlistedCamperDTO> {
+    let newWaitlistedCamper: WaitlistedCamper;
+    let existingCamp: Camp | null;
+
+    try {
+      newWaitlistedCamper = await MgWaitlistedCamper.create({
+        firstName: waitlistedCamper.firstName,
+        lastName: waitlistedCamper.lastName,
+        age: waitlistedCamper.age,
+        contactName: waitlistedCamper.contactName,
+        contactEmail: waitlistedCamper.contactEmail,
+        contactNumber: waitlistedCamper.contactNumber,
+        camp: waitlistedCamper.camp,
+      });
+
+      try {
+        existingCamp = await MgCamp.findByIdAndUpdate(
+          waitlistedCamper.camp,
+          {
+            $push: { waitlistedCampers: newWaitlistedCamper.id },
+          },
+          { runValidators: true },
+        );
+
+        if (!existingCamp) {
+          throw new Error(`Camp ${waitlistedCamper.camp} not found.`);
+        }
+      } catch (mongoDbError: unknown) {
+        // rollback user creation
+        try {
+          const deletedWaitlistedCamper: WaitlistedCamper | null = await MgWaitlistedCamper.findByIdAndDelete(
+            newWaitlistedCamper.id,
+          );
+
+          if (!deletedWaitlistedCamper) {
+            throw new Error(
+              `Waitlisted Camper ${newWaitlistedCamper.id} not found.`,
+            );
+          }
+        } catch (rollbackDbError) {
+          const errorMessage = [
+            "Failed to rollback MongoDB waitlisted camper creation after update camp failure. Reason =",
+            getErrorMessage(rollbackDbError),
+            "MongoDB camper id that could not be deleted =",
+            newWaitlistedCamper.id,
+          ];
+          Logger.error(errorMessage.join(" "));
+        }
+
+        throw mongoDbError;
+      }
+    } catch (error: unknown) {
+      Logger.error(
+        `Failed to create waitlisted camper. Reason: ${getErrorMessage(error)}`,
+      );
+      throw error;
+    }
+
+    return {
+      id: newWaitlistedCamper.id,
+      firstName: waitlistedCamper.firstName,
+      lastName: waitlistedCamper.lastName,
+      age: waitlistedCamper.age,
+      contactName: waitlistedCamper.contactName,
+      contactEmail: waitlistedCamper.contactEmail,
+      contactNumber: waitlistedCamper.contactNumber,
+      camp: waitlistedCamper.camp,
+    };
   }
 }
