--- backend/typescript/services/interfaces/waitlistService.ts
@@ -1,15 +1,0 @@
-import { WaitlistedCamperDTO, CreateWaitlistedCamperDTO } from "../../types";
-
-interface IWaitlistService {
-  /**
-   * Create a waitlisted camper
-   * @param waitlistedCamper the waitlisted camper to be created
-   * @returns a WaitlistedCamperDTO with the created waitlisted camper's information
-   * @throws Error if waitlisted camper creation fails
-   */
-  createWaitlistedCamper(
-    waitlistedCamper: CreateWaitlistedCamperDTO,
-  ): Promise<WaitlistedCamperDTO>;
-}
-
-export default IWaitlistService;
